refactor: migrate MeetingVisibility and MeetingType enums to shared enums folder
Move MeetingVisibility and MeetingType enums from interfaces to the proper shared enums folder structure for better organization and consistency.
